requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-telegram-bot[job-queue]==22.5